+++ modules/portal/portal-search-elasticsearch/src/com/liferay/portal/search/elasticsearch/internal/index/CompanyIndexFactory.java
@@ -20,7 +20,7 @@ import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.search.elasticsearch.index.IndexFactory;
-import com.liferay.portal.search.elasticsearch.util.LogUtil;
+import com.liferay.portal.search.elasticsearch.internal.util.LogUtil;
 
 import java.util.HashMap;
 import java.util.Map;
